Implement Reviewers corrections
