+++ src/modules/save.js
@@ -1,0 +1,5 @@
+const saveTodos = (tasks) => {
+  localStorage.setItem('tasks', JSON.stringify(tasks));
+};
+
+export default saveTodos;
